+++ udigest/tests/derive_compile_fails.rs
@@ -1,0 +1,5 @@
+#[test]
+fn trybuild() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("./tests/derive_compile_fails/*.rs");
+}